Replace int with uint8_t in most places

+++ libraries/FlightSimBreakoutBoard/src/FlightSimBreakoutBoard.cpp
@@ -24,30 +24,30 @@ void FlightSimBreakoutBoard::setup(const FlightSimBreakoutBoardConfig &config) {
   turnOffAllLEDs();
 }
 
-void FlightSimBreakoutBoard::setLED(int led_id, uint8_t value) {
-  const int expander_number = led_id / NUM_PINS_PER_IO_EXPANDER;
-  const int pin_number = led_id % NUM_PINS_PER_IO_EXPANDER;
+void FlightSimBreakoutBoard::setLED(uint8_t led_id, uint8_t value) {
+  const uint8_t expander_number = led_id / NUM_PINS_PER_IO_EXPANDER;
+  const uint8_t pin_number = led_id % NUM_PINS_PER_IO_EXPANDER;
   // LEDs are ordered from last pin back, so reverse order for indexing.
-  const int expander_id = num_digital_io_expanders_ - 1 - expander_number;
-  const int pin_id = NUM_PINS_PER_IO_EXPANDER - 1 - pin_number;
+  const uint8_t expander_id = num_digital_io_expanders_ - 1 - expander_number;
+  const uint8_t pin_id = NUM_PINS_PER_IO_EXPANDER - 1 - pin_number;
 
   digital_io_expanders_[expander_id].set_led(pin_id, value);
 }
 
 void FlightSimBreakoutBoard::turnOffAllLEDs() {
-  for (int i = 0; i < num_digital_io_expanders_; i++) {
+  for (uint8_t i = 0; i < num_digital_io_expanders_; i++) {
     digital_io_expanders_[i].set_all_led(0);
   }
 }
 
 void FlightSimBreakoutBoard::turnOnAllLEDs() {
-  for (int i = 0; i < num_digital_io_expanders_; i++) {
+  for (uint8_t i = 0; i < num_digital_io_expanders_; i++) {
     digital_io_expanders_[i].set_all_led(1);
   }
 }
 
 void FlightSimBreakoutBoard::loop(const FlightSimBreakoutBoardConfig &config) {
-  for (int i = 0; i < num_digital_io_expanders_; i++) {
+  for (uint8_t i = 0; i < num_digital_io_expanders_; i++) {
     digital_io_expanders_[i].sendStateUpdateOnButtonChange(joystick_);
   }
   analog_multiplexer_.sendStateUpdateOnAnalogInputChange(
@@ -57,13 +57,13 @@ void FlightSimBreakoutBoard::loop(const FlightSimBreakoutBoardConfig &config) {
 void FlightSimBreakoutBoard::setupDigitalIOExpanders(
     const DigitalIOExpandersConfig &config) {
   num_digital_io_expanders_ = config.num_expanders;
-  int joystick_start_index_for_current_expander = 0;
+  uint8_t joystick_start_index_for_current_expander = 0;
 
   // Iterate backwards through the expanders count to allocate the number of
   // LEDs on each expander.
-  int num_leds_per_expander[8] = {0};
-  int num_total_leds_to_allocate = config.num_leds;
-  int backwards_idx = num_digital_io_expanders_ - 1;
+  uint8_t num_leds_per_expander[8] = {0};
+  uint8_t num_total_leds_to_allocate = config.num_leds;
+  uint8_t backwards_idx = num_digital_io_expanders_ - 1;
   while (num_total_leds_to_allocate > 0) {
     num_leds_per_expander[backwards_idx] =
         min(NUM_PINS_PER_IO_EXPANDER, num_total_leds_to_allocate);
@@ -71,7 +71,7 @@ void FlightSimBreakoutBoard::setupDigitalIOExpanders(
     backwards_idx--;
   }
 
-  for (int i = 0; i < num_digital_io_expanders_; i++) {
+  for (uint8_t i = 0; i < num_digital_io_expanders_; i++) {
     digital_io_expanders_[i] = DigitalIOExpander();
     digital_io_expanders_[i].setup(config.expander_i2c_addresses[i],
                                    num_leds_per_expander[i],
