+++ libraries/FlightSimBreakoutBoard/src/subcomponents/DigitalIOExpander.cpp
@@ -2,8 +2,8 @@
 
 DigitalIOExpander::DigitalIOExpander() : gpio_board_() {}
 
-void DigitalIOExpander::setup(uint8_t address, int num_leds,
-                              int joystick_start_index) {
+void DigitalIOExpander::setup(uint8_t address, uint8_t num_leds,
+                              uint8_t joystick_start_index) {
   gpio_board_.setAddress(address);
 
   num_led_outputs_ = num_leds;
@@ -16,7 +16,7 @@ void DigitalIOExpander::setup(uint8_t address, int num_leds,
   for LED outputs).
   */
   uint16_t button_mask = 0U;
-  for (int i = 0; i < NUM_PINS_PER_IO_EXPANDER; i++) {
+  for (uint8_t i = 0; i < NUM_PINS_PER_IO_EXPANDER; i++) {
     // LED pins are assigned to the end of the pin mask.
     bool is_led = NUM_PINS_PER_IO_EXPANDER - i <= num_led_outputs_;
     bitWrite(button_mask, i, is_led ? 0 : 1);
@@ -42,7 +42,7 @@ void DigitalIOExpander::set_all_led(uint8_t value) {
 void DigitalIOExpander::readAndSetAllButtons(Joystick_ &joystick) {
   const uint16_t pin_values = gpio_board_.readButton16();
 
-  for (int i = 0; i < NUM_PINS_PER_IO_EXPANDER - num_led_outputs_; i++) {
+  for (uint8_t i = 0; i < NUM_PINS_PER_IO_EXPANDER - num_led_outputs_; i++) {
     last_button_state_[i] = bitRead(pin_values, i);
     joystick.setButton(joystick_start_index_ + i, last_button_state_[i]);
   }
@@ -51,8 +51,8 @@ void DigitalIOExpander::readAndSetAllButtons(Joystick_ &joystick) {
 void DigitalIOExpander::sendStateUpdateOnButtonChange(Joystick_ &joystick) {
   const uint16_t pin_values = gpio_board_.readButton16();
 
-  for (int i = 0; i < NUM_PINS_PER_IO_EXPANDER - num_led_outputs_; i++) {
-    const int current_button_state = bitRead(pin_values, i);
+  for (uint8_t i = 0; i < NUM_PINS_PER_IO_EXPANDER - num_led_outputs_; i++) {
+    const uint8_t current_button_state = bitRead(pin_values, i);
     // Only send update if button state has changed.
     if (current_button_state != last_button_state_[i]) {
       // Send update immediately upon button state change.
